sunday commit

+++ src/app/components/signup/signup.component.ts
@@ -1,16 +1,17 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-signup',
   standalone: true,
   imports: [FormsModule, CommonModule],
   templateUrl: './signup.component.html',
-  styleUrl: './signup.component.scss'
+  styleUrl: './signup.component.scss',
 })
 export class SignupComponent {
-  [x: string]: any;
+
   public userInfo: Array<{
     email: string;
     firstname: string;
@@ -23,22 +24,27 @@ export class SignupComponent {
     lastname: '',
     password: '',
   };
-}
+  public displayForm: boolean = false; // Make sure to declare displayForm
 
-adduser=()=> {
-  if (this.newUser.email && this.newUser.firstname && this.newUser.lastname && this.newUser.password
-  ) {
-    this.userInfo.push(this.newUser);
-    localStorage.setItem('userInfo', JSON.stringify(this.userInfo));
-    this.newUser = {
-      username: '',
-      email: '',
-      course: '',
-      location: '',
-    };
+  constructor(private router: Router) {} // Inject the Router
+
+  addUsers() {
+    if (this.newUser.email && this.newUser.firstname && this.newUser.lastname && this.newUser.password) {
+      // Check if email already exists in userInfo array
+      const emailExists = this.userInfo.some(user => user.email === this.newUser.email);
+
+      if (!emailExists) {
+        this.userInfo.push(this.newUser);
+        localStorage.setItem('userInfo', JSON.stringify(this.userInfo));
+        this.newUser = { email: '', firstname: '', lastname: '', password: '' };
+        this.displayForm = false;
+        this.router.navigate(['login']);
+      } else {
+        // Handle the case when email already exists
+        // For example, show an error message to the user
+        alert('Email already exists. Please use a different email.');
+      }
+    }
   }
-  this.displayForm = false;
-  console.log(this.userInfo);
 
-    this.router.navigate(['nav'])
 }
